Restructure PII processor (#11640)
* Restructure PII processor

* Restructure PII processor

* Format
